Fix voting restriction and name edit vulnerability, and remove temp files.
Fixed a bug where a user on a given IP address was only able to vote once
on the entire site, and not once per name. Also, a vote on a name, whether up
or down, could not be undone with the opposite vote. Added a bug somewhere in
the program as an exercise.
Next, fixed a vulnerability where the route for the edit prompt was left open,
such that any user could access it by entering the URL.
Finally, removed temp files ending in ~ from the repository, and modified
.gitignore to keep them from coming back.